extract the ZipEntryFileVisitor from childEntries in its own class
